requirements.txt
gym==0.26.0
numpy==1.25.2
pygame==2.1.0
PyYAML==6.0.1
matplotlib==3.8.2
pandas==2.1.4
scipy==1.11.4
